Add smart fit option
Using smart fit, the image will be scaled to fit, as long as the image
pixels are not scaled by more than a preset amount (default is 1.5px),
otherwise they will be surrounded by whitespace. There is also a
threshold value for this behaviour, so that you can avoid images being
almost stretched to the full width, default is 0.75, meaning that if
the image is 75% stretched, it will be set to cover the frame.